+++ src/utils/constants.ts
@@ -1,0 +1,10 @@
+export const ROUTES = {
+    login: {
+      name: 'login',
+      path: '/login',
+    },
+    forgot_password: {
+      name: 'forgot_password',
+      path: '/forgot_password',
+    },
+}
